Created the outline Projects, CV, and Contact pages
Created the projects, CV, and contact pages in an outline form.

Currently they are all identical to the homepage, bar the intro text. They also contain a unique heading (e.g. "Dan Harris | Projects"), instead of just "Dan Harris" on the homepage.

Also added comment sections for each of the pages in the styles.css document.

Unsure whether separating the CSS into multiple files will be best, or if keeping it as one file for optimisation reasons would be better.