Created a JS script that creates project cards automatically
Created a JS script that automatically creates new project cards with content.

This will later be expanded to create these automatically from a JSON or XML file, which will contain the content required, but nothing more.

Also improved the project card's structure and made it consistent between the test and the standard cards