Began to implement reading project data from a a JSON file
Started to implement reading the project card data from a JSON file and creating a unique project card from that data.

Currently it creates all of the necessary elements, but the contents are all undefined.

I am also cheating with this as the two methods written (one commented), aren't actually reading from a JSON, they are simply reading from a "fake" JSON file (really just a JS file with an object), and a JS object in the style of a JSON.

Actually reading from a local JSON file will require either node.js or JQuery.