Initial code for attempts to automate the project page
Initial commit for both a JQuery and node.js attempt to read from a json and automatically display the information on the projects page as separate projects.

Currently JQuery solution creates the correct elements, but populates the content with "undefined".

Node.js is the code (ProjectCardPopulator) that should read the projects json, however I have yet to create the node.js backend to implement it.