Changed how loading the projects from a json works
RemovedProjectCardPopulator.js  as I came to the conclusion that using node.js for loading a JSON file is overkill. In addition to node.js not playing nice with github pages (planned hosting service - only static pages).

Also changed projects.json to not only add another test project, but to also have it properly formatted as a JSON file.

Additionally commented out the JQuery code (as realistically JQuery is overkill for this task).

As a result I copy and pasted the data from projects.json into the js file to test whether the code for displaying the data once I have got it functions correctly