Projects are now dynamically added to the project page
Projects stored in Projects.json are now automatically added to the projects page.

To accomplish this, I have moved the webpages out from their own subfolder to be in the root folder.

Also deleted the projectsTEST file as it is no longer need.

As a result of the movement of the webpages in the folder structure I have updated the links to resources in the HTML so they are displayed correctly.

Altered the fetch code to only run when ok message received, as well as removing the JQuery code from the js as well as the jquery-min file.

NOTE: In order for the projects to be loaded locally, a local webserver must be run (currently utilising "live-server").