Minor changes to CSS and contact page removal from main css
Made some minor changes to the CSS to remove some extra padding, etc. and hid social links on mobile

Also removed all of the contact page css from the main stylesheet into its own style sheet. Currently not planning on using the page, but if I decide to implement the backend for it later it will still have all of its styling (also fixed a broken image link on the page)