Changed how padding is applied and improved layout on mobile
Changed how padding is applied to allow for background colours to span the width of the page (rather than just the div of the content).

Also (hopefully) improved some aspects of viewing the content on a mobile device. By drastically reducing the size of the padding and by changing the size of header text