Adding modal content to the page
Have quickly added dynamically adding modal content to the page.

Currently all of the required content is there, but there is an issue with the formatting of the close button.

Will need to link up the details buttons (or the cards themselves) to their respective modals. Will need to change the script for this

Also will have to change the script to allow the close buttons of all modals to work