Implemented a more flexible and feature rich modal content manager
Implemented a more flexible and feature rich way of populating a modals content.

Each modal is broken up into multiple sections, with each section consisting of a heading (h3), some text, an image and / or a video. Each section can consist of some or all of these elements.

Also updated the Modals.json file to reflect the new json structure