Updated the iframes to video tags
Updated the iframes to video tags where the video is not an embedded youtube link, as it was causing the website to try and download the videos on the first page load rather than embed them.

Changing the tags to videos has seemed to fix this.

Also updated the CSS and CSS class names to reflect this change.

Fixed some of the formatting in the HTML for the videos, but as its auto generated, its not really worth going any further